% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{add_aqli_color_scale_buckets}
\alias{add_aqli_color_scale_buckets}
\title{Add AQLI color scale buckets for all AQLI scales to a dataset}
\usage{
add_aqli_color_scale_buckets(df, scale_type = "pollution", col_name)
}
\arguments{
\item{df}{The dataframe to which the AQLI color scales will be added as a character column}

\item{scale_type}{This is a parameter of type \code{character} and can take one of 4 values: (a) \code{lyl}: AQLI life years lost (single year);
                  (b) \code{pollution}: AQLI pollution (in µg/m³, single year); (c) \code{lyldiff}:
                  Life years lost scale (for difference between 2 years); (d) \code{poldiff}: pollution
                  difference between 2 years scale (in %).

                  Note that, while plotting the actual map
                  that will use this function, make sure to add in the \code{fill} aesthetic of the mapping
                  function the following (defined here for each scale): \code{lyl} scale:
                   \code{forcats::fct_reorder(lyl_bucket, order_lyl_bucket)}; \code{pollution} scale:
                   \code{forcats::fct_reorder(pol_bucket, order_pol_bucket)}; \code{lyldiff} scale:
                   \code{forcats::fct_reorder(lyldiff_bucket, order_lyldiff_bucket)}; \code{poldiff} scale:
                   \code{forcats::fct_reorder(poldiff_bucket, order_poldiff_bucket)}}

\item{col_name}{the column name (type:\code{character}) that will be used to map pollution/life years lost (whether
single year or difference between 2 years) values to the AQLI colors. Note, that
this column should be present in \code{df}.}
}
\value{
returns a data frame with an additional column (of class "character") that has the buckets
        corresponding to the \code{scale_type} specified.
}
\description{
Adds the buckets (mappping AQLI colors to the corresponding pollution or life years lost
value range buckets) as a column to the underlying dataset (depending on the underlying scale in use).
The dataset should have a column that will be used for mapping colors to values.
This function comes in handy while plotting graphs that need to be colored according to AQLI color scales.
}
\examples{
df \%>\% add_aqli_color_scale_buckets(scale_type = "pollution", col_name = "pm2021")
df \%>\% add_aqli_color_scale_buckets(scale_type = "lyl", col_name = "llpp_who_2021")
df \%>\% add_aqli_color_scale_buckets(scale_type = "poldiff", col_name = "pm_example_pol_diff_col")
df \%>\% add_aqli_color_scale_buckets(scale_type = "lyldiff", col_name = "llpp_example_lyl_diff_col")

}
