% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{aqli_bar}
\alias{aqli_bar}
\title{Plot aqli pollution and life years lost bar graphs}
\usage{
aqli_bar(
  df,
  scale_type = "pollution",
  x_var,
  y_var,
  title,
  subtitle,
  x_label,
  y_label,
  legend_title,
  caption
)
}
\arguments{
\item{df}{AQLI data that will be used to plot the bar graph.}

\item{scale_type}{One of \code{pollution} or \code{lyl}. Depends on the underlying underlying value of \code{y_var} column  (from \code{df}) that is plotted.}

\item{x_var}{A column of type \code{character}, e.g. depending on the bar graph, this could be country, state, district. See examples below. Note
that this will end up being plotted on y axis as the function by default flips the axis using \code{coord_flip}. You can flip it back
by piping the output of this function again to \code{coord_flip}.}

\item{y_var}{A column of type \code{numeric}, e.g. life years lost or pollution in a given year. Note that this will end up being plotted on y axis as the function by default flips the axis using \code{coord_flip}. You can flip it back
by piping the output of this function again to \code{coord_flip}.}

\item{title}{The title of the bar graph.}

\item{subtitle}{The subtitle of the bar graph.}

\item{x_label}{x-axis label of the graph.}

\item{y_label}{y-axis label of the graph.}

\item{legend_title}{Title for the legend of the bar graph.}

\item{caption}{Caption for the bar graph.}
}
\value{
returns a bar plot given the above specifications.
}
\description{
Plots pollution and life years lost regional bar graphs in AQLI colors.
}
\note{
Please note that the aqli color buckets should already be added to \code{df} using \code{add_aqli_color_scale_buckets} function before it gets
feeded to \code{aqli_bar}
}
\examples{

# Plotting the graph of life years lost relative to WHO guideline (llpp_who_2021) for all prefectures (name_2) of a aqli gadm2 dataset
# that is filtered for China.

df \%>\%
aqli_bar(scale_type = "lyl", x_var = "name_2", y_var = "llpp_who_2021", title = "", subtitle = "", x_label = "Prefecture", y_label = "Potential Gain in Life Expectancy (Years)", legend_title = "Potential gain in life expectancy (Years)", caption = "") +
 theme(plot.background = element_rect(fill = "white", color = "white"))

}
