% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{aqli_hist}
\alias{aqli_hist}
\title{Plot a histogram of pollution/life years lost columns in the AQLI dataset}
\usage{
aqli_hist(
  df,
  scale_type = "pollution",
  col_name = "pm2021",
  region_name = "enter region name"
)
}
\arguments{
\item{df}{An AQLI dataframe that contains the columns that will be used to plot the histogram.}

\item{scale_type}{This can take one of 2 values \code{"pollution"} or \code{"lyl"}.}

\item{col_name}{The column name (in quotes) from \code{df} for which the histogram is to be plotted.}

\item{region_name}{Plot subtitle (in quotes)}
}
\value{
returns a histogram plot based on the specifications provided.
}
\description{
Plots a histogram of either pollution or life years lost values in AQLI color scale
}
