% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{aqli_map_sentence_stats}
\alias{aqli_map_sentence_stats}
\title{AQLI website map summary sentence cross-check function}
\usage{
aqli_map_sentence_stats(
  df,
  level = "global",
  data_type = "lyl",
  year1_col = "llpp_who_2021",
  year2_col = "llpp_who_2014",
  country_name = "India",
  state_name = "Uttar Pradesh",
  district_name = "Ghaziabad"
)
}
\description{
AQLI website map summary sentence cross-check function
}
