% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{gadm_level_summary}
\alias{gadm_level_summary}
\title{Collapse AQLIs most granular processed data to less granular levels}
\usage{
gadm_level_summary(df, level_col_name_vec, years, perc_red_by)
}
\arguments{
\item{df}{AQLI processed gadm level 2 dataset, corresponding to the data-dictionary at this link (to be added).}

\item{level_col_name_vec}{a vector specifying the highest resolution at which the \code{df} is to be collapsed. This
vector is to be specified starting from lowest to highest resolution (for a non-continent summary).
The four available resolutions are: \code{continent}, \code{country}, \code{name_1}, \code{name_2}.
For a \code{continent} level summary, the vector would look like: c("continent"). But, for all
"non-continent summary" the vector would be specified from lowest to highest resolution. For example -
for a state (\code{gadm1}) level summary, the vector would look like: c("country", "name_1"). See examples
below for more use cases.}

\item{years}{a vector of years for which the summary data is needed, e.g. \code{c(1998, 2000, 2021)}.
The 2023 AQLI dataset provides data on the following years 1998 to 2021.}

\item{perc_red_by}{custom reduction percentage (numeric), i.e. if columns for custom reduction are needed. For example, if we want to see
pollution (and corresponding lyl) columns for a 10% reduction relative to the years specified, then this
parameter would be 10.}
}
\value{
returns a collapsed dataset with
}
\description{
AQLI (in 2021) reports annual average PM2.5 and life years lost at 3 levels:
gadm0, 1, 2 (\href{https://gadm.org/}{gadm}). These levels define the administrative region boundaries of a
particular "country". gadm0 corresponds to a country. gadm1 corresponds to
"state" and gadm2 corresponds to "district/prefecture/county" etc. This function
takes in a gadm2 dataset and has the functionality to collapsing it gadm1 and
gadm0 (and gadm2, which is equivalent to the original data).
Furthermore, one can specify the \code{years} to specify the years for which
the collapsed data is needed and \code{perc_red_by} (percentage reduction by) to specify if
additional columns showing pollution levels and life years lost after a custom reduction are to be also c
computed
}
