% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{plot_aqli_pol_lyl_map}
\alias{plot_aqli_pol_lyl_map}
\title{Plot AQLI and life expectancy gains single year maps}
\usage{
plot_aqli_pol_lyl_map(
  gadm2_aqli_csv,
  gadm2_aqli_shapefile,
  gadm1_aqli_shapefile,
  gadm0_aqli_shapefile,
  region_level,
  col_name_plt,
  plot_type,
  country_name,
  state_name,
  district_name
)
}
\arguments{
\item{gadm2_aqli_csv}{A CSV file containing AQLI and life expectancy data at the GADM2 (district/county/prefecture) level.}

\item{gadm2_aqli_shapefile}{A geographic shapefile for GADM2 (district/county etc) administrative level polygons.}

\item{gadm1_aqli_shapefile}{A geographic shapefile for GADM1 (state/provinces) administrative level polygons.}

\item{gadm0_aqli_shapefile}{A geographic shapefile for GADM0 (country) administrative level polygons.}

\item{region_level}{The administrative level of the region to plot (country, state, or district).}

\item{col_name_plt}{The name of the column in the CSV containing pollution or potential life expectancy gains data.}

\item{plot_type}{The type of plot: "lyl" for life expectancy or "pollution" for AQI.}

\item{country_name}{The name of the country for which to generate the map.}

\item{state_name}{The name of the state (if applicable) for which to generate the map.}

\item{district_name}{The name of the district (if applicable) for which to generate the map.}
}
\value{
A ggplot object displaying the map.
}
\description{
This function takes in AQLI (Air Quality Life Index) and life expectancy data in CSV format, a
and geographic shapefiles representing different administrative levels. It generates a map visualizing
either the potential gain in life expectancy or air pollution levels (PM2.5) for a specific region
(country, state, or district) based on the specified parameters.
}
