% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{plot_country_level_gbd}
\alias{plot_country_level_gbd}
\title{Plot Country Level GBD Data}
\usage{
plot_country_level_gbd(
  gbd_master_data,
  country_name = "India",
  degree = "most",
  n_threats = 10
)
}
\arguments{
\item{gbd_master_data}{The GBD master dataset for all countries.}

\item{country_name}{The name of the country to plot the GBD data for. Default is "India".}

\item{degree}{Specify whether to consider the most or least deadly threats. Default is "most", other option is "least".}

\item{n_threats}{The number of most or least deadly threats to consider. Default is 10.}
}
\value{
A plot representing the GBD data for the specified country and threats.
}
\description{
Plots the GBD data for a specific country, filtering based on the number of most/least deadly threats to life expectancy.
}
