% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\docType{data}
\name{themes_aqli_base}
\alias{themes_aqli_base}
\title{Add AQLI base custom theme to plots}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
themes_aqli_base
}
\arguments{
\item{""}{no parameters needed. See examples for usage.}
}
\value{
when added to a plot, returns a plot with the AQLI plots base theme.
}
\description{
Adds a custom AQLI theme to plots in a single command
}
\keyword{datasets}
