% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqli.processed.data.analysis.helper.R
\name{trendlines_aqli}
\alias{trendlines_aqli}
\title{Plot AQLI pollution time series plots (trendlines)}
\usage{
trendlines_aqli(
  gadm2_file,
  level = "country",
  country_name = "India",
  state_name = "NCT of Delhi",
  district_name = "NCT of Delhi",
  start_year = 1998,
  end_year = 2021
)
}
\arguments{
\item{gadm2_file}{aqli gadm2 (district/county/prefecture level) master file.}

\item{level}{one of \code{country}, \code{state} or \code{district}.}

\item{country_name}{country for which (the country itself, or a region within that country) the trendlines are to be plotted. Defaults to "India".}

\item{state_name}{state within the \code{country} specified for which (the state itself, or a region within that state) the
trendlines are to be plotted. Defaults to "NCT of Delhi"}

\item{district_name}{district within, the \code{state}, which is in \code{country} specified for which the
trendlines are to be plotted. Defaults to "NCT of Delhi".}

\item{start_year}{year from which the trendline will start (defaults to 1998).}

\item{end_year}{year at which the trendline will end (defaults to 2021).}
}
\value{
a trend lines plot from \code{start_year} to \code{end_year} for the region specificed at the level specified in \code{level}
}
\description{
Plots pollution trends at country, state and district levels
from a given start year to a given end year
}
